operator: add kube_cloud_config controller
The kube_cloud_config controller

- performs transformations on the user provided cloud config to create a new cloud config that
  stitches in the information from infrastructure.config.openshift.io object
- it creates and manages a stitched cloud config config-map and updates the status of the infra object
- it provides a way to have per platform transformers for stitching cloud config

There are 2 transformers for now,
- as-is
This copies all the contents of the source config map to target
This is used when there are no special transformers and the spec one needs to copied as-is

- aws
This transformer uses the service endpoints set in the infra.status.platformStatus.aws.serviceEndpoints to add
ServiceOverrides sections
If both user provided cloud-config and the infra have service endpoints, it returns an error

There is no Marshaller for gcfg used by the aws kube cloud-provided codebase, therefore we create ServiceOverrides sections using text/template and append those to the preexiting cloud config data.
It copies all the other keys in the config map as-is and only modifies the cloud config containing key.