kube_cloud_config: support for other Azure cloud environments
Populate the `cloud` field in kube cloud.conf stored in the kube-cloud-config
ConfigMap in the openshift-config-managed namespace. This field informs in
which Azure cloud environment the cluster is installed and consequently which
Azure API endpoint should be used when communicating via the Azure SDK.

The value populated comes from the `.status.platformStatus.azure.cloudName`
field of the infrastructure.config.openshift.io resource. If the field is
empty or missing, then the default "AzurePublicCloud" value is set for the
`cloud` field in the kube cloud config. If the field generated from the
infrastructure resource conflicts with the field in the user-provided kube
cloud config, then the controller will error on syncing the infrastructure
resource.

See https://github.com/openshift/enhancements/pull/321

https://issues.redhat.com/browse/CORS-1444