aws_platform_service_location: resync when user unsets the service endpoints in spec
Previously when the user would reset the service endpoints in the spec, the controller would exit early as no work needed to be done.
That means it would leave the status set with previous service endpoints.

Now the controller resets the service endpoints when the user set the service endpoints to empty.