update AWS platform status migration controller for all platforms
The AWS platform status migration controller was created to populate
status.platformStatus in the infrastructure resource for AWS clusters
created in 4.1, prior to the existence of that field. One of the fields
that the controller sets is the status.platformStatus.type field, which
is set to the type of the platform--in this case AWS. Other operators
are looking to that field to determine which platform the cluster is
running on. However, for non-AWS clusters created in 4.1, the field
remains blank.

With these changes, the AWS platform status migration controller has
been re-purposed to fill out the appropriate parts of status.platformStatus
for all platforms, not just for AWS. The only other platform used in 4.1
besides AWS was None, and there are no fields in the infrastructure resource.
Therefore, for all platforms other than AWS, the only field that the
controller populates is the status.platformStatus.type field.

https://bugzilla.redhat.com/show_bug.cgi?id=1890038